Functionlaize code and test groupme calls
